+++ src/Svelte/Search/BooleanCriterion.svelte
@@ -1,4 +1,6 @@
 <script>
+    import { _ } from "svelte-intl"
+
     /**
      * @type {string} - Display name of the filter
      * @prop
@@ -17,8 +19,8 @@
 <label class="filter" class:empty="{value === '-1'}">
     <span>{field}</span>
     <select bind:value>
-        <option value="-1">All</option>
-        <option value="1">Yes</option>
-        <option value="0">No</option>
+        <option value="-1">{$_("filter.any")}</option>
+        <option value="1">{$_("filter.yes")}</option>
+        <option value="0">{$_("filter.no")}</option>
     </select>
 </label>