+++ src/Svelte/Search/ListCriterion.svelte
@@ -1,4 +1,6 @@
 <script>
+    import { _ } from "svelte-intl"
+
     /**
      * @type {string} - Display name of the filter
      * @prop
@@ -19,7 +21,7 @@
 <label class="filter" class:empty="{value === undefined || value === ''}">
     <span>{field}</span>
     <select bind:value>
-        <option value="">All</option>
+        <option value="">{$_("filter.any")}</option>
         {#each completions as completion}
         <option>{completion}</option>
         {/each}